rust json-ast: typed AST and update
