+++ tests/json-ast/x/rs/cross_join.rs.json
@@ -1,5 +1,5 @@
 {
-  "file": {
+  "root": {
     "kind": "source_file",
     "start": 0,
     "end": 2561,
